* commit '3ee774ab5296ddc0551d0f125b077fe363214345':
  Fixed a useless cast in src/pancake/SeedDBWriter.cpp.

+++ src/pancake/SeedDBWriter.cpp
@@ -127,8 +127,7 @@ void SeedDBWriter::WriteSeeds(const std::vector<PacBio::Pancake::FastaSequenceCa
     }
 
     for (size_t i = 0; i < seqs.size(); ++i) {
-        WriteSeeds(seqs[i].Name(), static_cast<int32_t>(seqs[i].Id()),
-                   static_cast<int32_t>(seqs[i].size()), seeds[i]);
+        WriteSeeds(seqs[i].Name(), seqs[i].Id(), seqs[i].size(), seeds[i]);
     }
 }
